support for searching datasets for known data formats it is now possible to specify an array based on a format and no length, and the app will look for that data based on known pointers. It will automatically match the longest values, and the results will be stored in the metadata file. Therefor, this search will only occur when loading a file that has no associated metadata.
This searching functionality is exposed in a new model subclass, so that the
tests won't be effected. However, the functionality exists in the general model,
so it's still possible to write specific tests based on bugs found.

Since an array can call something a string when it's not, added a new "ErrorPCS"
implementation of IDataFormat, to display the raw byte in red. This will help
find string errors and possibly missing string characters.