ThemeSelector uses a swatch
the theme selector was still a bit confusing, with both a light and dark mode. It's difficult (and not really neccissary) to get them both to look good at the same time. So now there's just one theme, the text is more clear, and I added swatches for selecting the primary text / background colors.