Show data instead of code where applicable
Not everything should be disassembled as code. If something tells me that it's loading a word from a given address (ldr), I know that the stuff there is data instead of code. So I should show it as such.

Not everything should be disassembled as code. If I read the code "b", "bx", "bl", or "pop pc", then I know that the bytes after it aren't code until something jumps to it. Go ahead and use this knowledge to stop showing code at that point.

Make it possible to scroll / copy from the code tool.