Make Items Expandable
If I intend to add more TMs, then I need to be able to add more items. Not only does this require updating the number of items, but I also have to update the number of item images (still never found the item images in Ruby/Sapphire...). But that's not all! We also need to update the code that refers to the number of items that we have. So here's a quick-edit aciton that updates code referring to the number of items.