Split TM/HM data
making TMs expandable involves making it easy to add new TMs to the items table. I don't want to force the user to make all TMs in a continuous set, and I don't want to require only a single gap or leaving the HMs where they are, like some other online tutorials suggest. Instead, I intend to update all of the TM/HM checking code to do it based on the item name. The item name will be TMxxx for TMs and HMxxx for HMs, for some length of numeric characters after the TM or HM header. I can use this knowledge to PARSE the number after the TM/HM header and get the TM/HM number that way.

Doing all of this requires that HMs be stored in a separate list from TMs. So that's the first change that I made: the 'MakeTMsExpandable' button now begins by splitting all the HMs and hmCompatibility data out into a separate table.