Patch item removal code to not care about TM50's itemID
testing revealed that when adding a TM51, the TM is not removed after use. I need to add patches to the item removal code. The code currently works by comparing the itemID against TM50. If the item ID is higher, then it's assumed to be an HM. That's... no longer correct. So now I have the code check to see if the itemID refers to a TM or HM using the IsItemTmHm function that we added to the game, which returns r1=2 if the item is an HM.