Refactor AutoSearchTests to make them run much faster
XUnit lets you run tests in parallel, but provides for "Test Collections", where tests within the same collection will never be run in parallel. This is a useful feature. However, by default, tests within the same class are always considered part of the same collection. This means that any test that involves actually loading a file and running the auto-search routine won't be run in parallel, which makes the tests in this class take a lot longer.

To fix this, I've made use of another XUnit feature: Test Fixtures. A test fixture lets me create a object that gets reused by all the tests in a class. So I can run the parallel code once in the fixture, and then have the never-modified collection available for use in the tests. So even though the tests don't run in parallel, the most expensive part (reading and auto-searching the files) is done in parallel, and only once.