Fix editing bug for Text Tool / level moves
There was a bug where editing level-moves via the tool wouldn't work if any of the moves contained spaces. Added a test to capture (See ViewPortEditTests), then fixed it (see changes to PLMRun).

Fixed a bug in my base test class.

ViewPort now informs the text tool immediately when a new run is created. So you don't have to click off and back on to get the text tool to update correctly now.