Thumb Bugfix
decompiled thumb code for "mov r7, r8" was showing up as "mov r15, r0" because I was getting the high bit wrong. Added a test that showed how it should be, and fixed it. But writing the reverse test for the compiler found a similar bug in the compiler. That's also fixed.