Cut in table should not delete entire table
Tables can contain many entries. If the user decides to delete or cut an entry, it should only effect that single entry instead of the entire array.

If the cut or delete has the entire table selected or also selects something from outside the table, the correct behavior is less clear, so keep the implementation the same.