Pointer menu item: Repoint to New Copy
Make it possible to duplicate data at an anchor. This new data can be edited independently of the original data, and will only be pointed to by the one location you chose. All other pointers will still point to the original version.

As part of this change, runs need the ability to duplicate themselves at a new starting point. Refactoring runs to add the new functionality.

As part of this change, the Model needs to expose its method of finding free space in the ROM.

Added a new test that shows the new feature.