Blinky Cursor for Hex Area
I finally did it. Multiple users asked for a blinking cursor to indicate editability in the Hex area. Well, now there's a blinky cursor.

Since the entire HexContent is designed to act as a single layer, and all the extra added effects live in TabView, I added the blinky-cursor code to TabView. It operates by drawing and erasing a tiny rectangle at the text insertion point, calculated using logic from FormatDrawer. We can't use a TextBox because a TextBox won't show its cursor when not focused, and we need the focus to stay on the HexContent.