First pass at adding code for rotated bit arrays
a user had the comment that bit arrays would be useful if they could show all the pokemon for a given TM, not just all the TMs for a given pokemon. The idea here is to add a bit-array element after a enum element if that enum is used by a bit array.