Include Stream links in table tool.
Some streams depend directly on tables. When working with the table tool in one of those tables, we can provide a reverse-lookup for any streams that use the element we're currently looking at. For example, if we've selected "charmander" in the pokestats table, we can do a reverse-lookup to see charmander's uses in the egg moves table.

This is a bit more manual than I'd like. Hopefully a better abstraction becomes apparent as I add more streams.

Also, in order for this to work, I need to be able to know the name of elements from text tables. As a bonus, this means that Ctrl+G->items/potion now works as expected.