Tool should auto-switch in some situations.
There are times where the table tool doesn't make sense but the text tool does, and vice-versa. In such situations, automatically switch from whichever of those tools is open to the other one.
Note that this will not automatically open a tool when none are open, nor will it mess with the code tool, nor will it change the tool in situations where both are applicable (such as the text portion of a table, or a run that is both a table and a stream). But it should reduce confusion and clicks in many common situations.