Don't auto-open tools
As useful as this bug was, it's preventing people from noticing the panel on the left. As the editor grows, this panel will become more useful. I'm going to remove the bug, so that the tool panel no longer auto-opens. This requires more work from the user, but that work causes them to discover the panels existence (hopefully).