Bugfix: don't change combo index on selection
For lists where multiple elements have the same text (example, the multiple 'picnicker' trainerclassname entries), loading the item with the table tool always selects the first element with that text in the combobox, even if that's not the index that was bound to. This is a binding issue.

To fix it, the ViewModel uses a slightly more complex data-type that wraps the string. The two options are now seen as different unequal objects, so no coercion happens.