Don't require table pointers to point to valid anchors
In many fangames, the game contains an error where a pointer in a table points to something unexpected.
* I need to allow the pointer, because the table format requires that it be a pointer, and that's how the game reads it.
* I need to allow the destination to not be an anchor, because the destination may be being used for something else.

To make sure I'm handling this correctly for a wide variety of cases, I added `BasicLoadTests.CanLoad` for a wide variety of romhacks.

Also related, I added another section to `PokemonModel.ResolveConflicts` that validates pointers within tables, instead of only pointerruns.