ViewPort's "Clear Format" context menu button should clear pointers too
Previously, right-clicking on an actual anchor would let you clear the format of the anchor but keep the anchor, while right-clicking on the format without the anchor would let you clear both the anchor and the format. Not only is this backwards, but it's inconsistent.

Now, a much simpler system: if you use the right-click to clear the format, the app assumes that the entire anchor was in error, so the anchor and any pointers to it are all cleared, but no data is changed.