Initial Pass at creating a general table-stream
Allow a table to point to a stream with a format specified as a table, with either a constant length, specific hex end code, or length based on another element of that table. This only works if the second format is "simple": this doesn't replace any of the exist stream formats.

plm: has a 7/9 bit split
egg: is either a pokemon or a move depending on the value
pcs: has escape sequences
tpt: has a variable format