Improve trainer team usability
Editing a pointer's stream element in the table shouldn't overwrite that stream element in the table. Overwriting the other controls is fine, but the self-overwrite makes the stream control very hard to use.

Modified logic around the Combobox tool segment to work even if the element segment isn't an EnumSegment. We DON'T expect this to happen, so we have an assert to catch it, but there's a latent bug somewhere causing this to happen. Until I can verify that the bug is fixed, I need to guard against it.