Show selected raw bytes in the bottom bar
A common feature request is to see the raw hex of data that is formatted, either to better understand what the editor is doing or because the user does not trust that the format is correct. Besides putting this in the raw tool, there's also room for it in the status bar. Added it in so the raw bytes are easily visible / copyable.