Improve child verification logic for table streams
The TableStreamRun/LengthFromParentStreamStrategy/GetCount has trouble getting the Count correctly during the initial verification, since the parent table hasn't actually been added to the model yet. So GetCount needs access to both the parent table's segments and the name of the pointer field. This lets the LengthFromParentStrategy to math to figure out segment offsets without needing the parent to actually be in the model. This extra info is only needed during verification: it's ok if it's `null` once the run is already added, such as during `CompleteEditOperation` and `TableTool`.

Add verification logic so we can know that we found a reasonable number of multichoice for each test rom. DarkRising's data is bugged, so ignore that one.