Check for unnamed anchors with no pointers
The system assumes that every anchor either has pointers to it or a name. An anchor with no poniters to it and no name isn't really allowed. So actually check to make sure we don't have any.

One way we could get this was removing a pointer to a pointer. The destination pointer would no longer have any sources to it, but since it still had a format, it was assumed to still be safe. This lead to pointers that had false anchors.