Update Tutor Expansion
There's a bug in tutor expansion: if you try to use tutor 16 (0x10) in a script, it no longer asks what pokemon you want but just tries to teach it to the first. This is incorrect.

While looking into this, go ahead and update the tutor code to be easier to work with in the future.