`MakeTutorsExpandable` updates tutor special (in FR/LG)
In FR/LG, there's a limiter built into special 397 (the move tutor special) that makes it behave differently after the element 14 (0-14 act one way, 15+ act a different way). I added a one-byte change to switch the branch condition to 'never'.

In order to do this, I needed to locate the special list, so I went ahead and added it to all the games, complete with a unit test to verify.

The specials table + Clover exposed a bug: when adding anchors from a table with an unknown pointer format, we need to clear the data of existing conflicting runs. Example: a text run that goes over the anchor isn't actually a text run.