Implement search for branch-link
in thumb code, branch-link commands are relative. That means that the instruction for the same destination will look different depending on the source.

As a user, I want to know what thumb code calls a routine I'm looking at. That's most likely going to be via a branch-link command, although it could be through a blx. Let the user use Find -> "bl <someaddress>" to search for branch-links that jump to a given location.