Unformatted Anchors can be cleared from right-click menu
A common user interaction is to notice that a table is too short and want to extend it. But doing so requires that the user manually clears any formats that are in the way of extension: such as anchors in that area from code that 'looks' like a pointer. To facilitate making this manual process easier, it should be possible to manually clear each false pointer without having to go to the pointer sources.

Basically, just let the user right-click and "Clear Format" even if the anchor has no format.