Performance Improvement: GetSelectedByteContents
GetSelectedByteContents was super slow, so I made 2 performance improvements.

(1) Cache off all of the bytes as strings. There's only 256 options, so storing them all isn't too hard, and this keeps us from having to recalculate them all the time.

(2) Update the SelectedBytes lazily. Instead of doing the work in `UpdateSelectedBytes`, have that function just set it to null. Then use the null check when `SelectedBytes.get` is called to actually do the work. This way, if no one asks for it, it's never actually calculated, saving a bunch of time in many situations.

There's no unit test because this is a performance change, not a behavior change.