Stop trying to scroll smartly on Undo
I've tried a bunch of different Heuristics for how to automatically scroll to the appropriate location when undo/redo is used. But none of the metrics have worked. Perhaps I'll try again later, but for now, let's just NOT scroll during undo/redo operations. It's not always the best thing to do, but it's extremely easy to grasp as a mental model.