Refactor to be able to support the 1.1 versions
Each of Ruby/Sapphire/FireRed/LeafGreen has a 1.1 version as well as the 1.0 version floating around online of the English GBA rom. I should be able to tell the difference between the two versions and load tables correctly.

This refactor makes it possible to tell the difference. Loading the tables correctly is coming next.