Add support for pokedex search tables
The tables refer to pokemon by their nationaldex number, not by their pokemon ID. So I needed a way to use an index-enum. That's the new test in NestedTableTests. That's also the reason for the change in StartCellEdit, ModelCacheScope, and IDataModel.

The new test in AutoSearchTests proves that the 4 new tables work correctly. That's also the reason for the change in StartScreen and tableReference.