Don't crash when loading TOML tables in the wrong order
If you try to load the tutorcompatibility table before the tutormoves table, loading the tutormoves table will change the length of the tutorcompatibility table. Under normal circumstances, this checks if the table needs to be expanded and if the data needs to be shifted around. But in a first-load scenario, we don't need (or want) to do any of that. So detect the first-load case and just skip out of the relocate/expand logic.