TableTool is open by default
To help users understand the features of the program, go ahead and have the table tool open by default. This has the benifit of showing a drop-down list with all the current known tables, so the user can quickly jump to them. This is similar to how other programs show some of their tools by default, helping the user know what things are possible with the application.

Also move the Table Tool to be the top tool for consistency.