ClearFormat to prevent double-format issues when expanding bitsegments during initialization
When loading from a TOML file, the bitsegment might get created before the table that it's based on. If so, then the bitsegment (and thus the entire table) will need to be expanded. When that happens, we need to expand in-place, because the original format was too short to encompass the data.

When that happens, we need to clear away whatever format is already in that spot, because that format is known to be wrong.

Related change: do a ResolveConflicts step as soon as we finish loading the metadata, to make sure that the metadata-load didn't leave us in a bad state.