Introduce the `MakeMovesExpandable` Quick Edit Item
Also adds automated tests to show that the code works like I expect. Note that this hasn't been tested in the actual games yet.

Note that a version of QuickEditItemDecorator needs to be checked in to avoid a boot-strap issue, since QuickEditItemDecorator is needed in the same assembly as IQuickEditItem. I could alternatively do a double-build of only the interfaces to fix this, but that seems heavy-handed for now.