Adjust how move expansion works
There are 5 pointers to the PP field in the movedata table. One of the goals of move-expansion is to simplify this by shifting these directly to the movedata table, so that repointing works automatically.
The 5 routines call only 2 subroutines using the PP field pointer, so my previous attempt changed those 2 subroutines. However, I failed to realize that (lots of) other code also calls these subroutines, so changing these subroutines has adverse side-effects. To correct this, I should change the 5 routines instead of the two subroutines.

This has been prototyped with FireRed and needs to be expanded to the other 8 roms.