Table Tool Pointer Editing Improvements
Previously, any invalid values typed into a pointer field in the table tool would be overwitten immediately with correct data. This made it almost impossible to manually type a value that would eventually be valid, such as `<null>`. This is now corrected: if the user types a value that would be an error, the error value is left in the field and no change or notification is made to the model.