Better bitfield editing support
With this change, you can now type a name of a bit into a bitfield to turn it on, and type a lone dash '-' to clear the bitfield. This makes it possible to have paste-scripts know the names of the moves they want to set, and then set them. For example:

`@tutormoves/bulbasaur - "mega kick" "solar beam" "grass knot"

The dash clears the bitfield, and then typing the names sets those bits.