Store the version in the metadata TOML file
This allows us to detect what version of the app was used to last save changes to this file, which lets us detect when the user has upgraded HMA and is now using it on a file last touched with an older version.

For example, we could notice that there were new auto-tables added since the last version, and the user might want us to add them.