Fix test `ExtendCompressedLzTokenShortensLzRun`
This test was failing (intentionally) in the previous version. Added logic to read the modified run after a data change via ViewPort interaction, and calculate the next compressed length. The code should also handle repointing in the case of extension, but that's not tested yet. It does, however, clear unused bytes to `FF` if the new compressed length is shorter.

I also separated the idea of a IDataFormatInstance from a IDataFormatStreamInstance.
* a `IDataFormatInstance` represents a single format that is multiple cells long. The entire format should be selected at once. Examples include a level-up token (2 cells) or a pointer token (4 cells).
* a `IDataFormatStreamInstance` represents a single-cell format that is part of a 'stream' that is multiple cells long. The format includes a reference to the start of the stream and the data format's index within the stream for the Text Tool. Examples include ASCII and PCS data. The user can freely type to edit the stream, but each format is only a single-cell long.

This isn't actually used yet, but the separation of the concepts should allow me to simplify ViewPort cell selection group logic in the future.