Support index-hints for sprites
Sometimes, the sprites don't each have a matching palette. Instead, there's a smaller number of palettes that are shared between all the sprites. In this case, it's useful for the sprite's palette hint to be able to be a name of a table who's length matches a table of palettes. Then the palette from that palette table can be deduced from the indexing element.