Bugfix: load palettes correctly after goto
The table tool is now in charge of fixing up all the sprites after it finishes filling the left-rail. This means that the sprite elements no longer need to work with only part of the left-rail loaded, so they can get their palette right every time.

As a side-effect, the TileImage now needs to potentially wait to load render, for when the width/height have been updated but the data hasn't yet.