Add annotation option
The script tip may seem less out-of-place if the code being tipped has some sort of visual cue, like a color change.

This implementation shows a proof-of-concept of how that might look by drawing a TextBlock directly over the existing text. But because of how text is rendered, it looks... terrible. A better thing to do would be to not draw the under-text, and draw the caret over both layers. This requires a more specialized control and could also be used for general syntax-highlighting. Revisit later.