Fix Castform sprite/palette pairing
when rendering a sprite, include an offset for which page of the sprite is being rendered. For a multi-page sprite with single page palettes or single-page sprites with multi-page palettes, this changes nothing. But for a multi-page sprite with multiple palettes, this allows the pages to correlate 1-to-1 as expected.