Fix app hang
LiquidCrystal has some pokemon teams that are 256 pokemon long. That was getting misread as 0 pokemon long, so it was causing a run of zero pokemon.
* Correctly read pokemon team counts as 4-bytes.
* If a run is length zero, that could hang the app, because several parts of the code loop through runs based on their length. So a length of zero doesn't move forward! These parts of the code are modified to always move forward at least one byte, even if a run is zero bytes long.
Note that a pokemon team could _still_ have a length of zero. It would appear as a warning in debug mode, but will no longer cause the app to hang.