Clear old sources from anchors when observing a run getting shorter
When a table gets shorter, some pointers leave it. These pointers are no longer part of a format, so they aren't pointing to anything. Update the destinations to know they're not being pointed at. Do this by doing a ClearFormat when we run into this case.
Because we're doing a ClearFormat, we need to make sure we do an ObserveAnchorWritten instead of just an ObserveRunWritten. In this case, the run coming in already knows what's pointing to it, so we don't need to search for pointers to it.