Refactor: share page add/remove code
Between LZSprites and LZPalettes, the code to add/remove pages is actually extremely similar. Let's add a common base class (PagedLZRun) so that they can actually share the code.

With the refactor complete, also fixing a few bugs with the add/remove CanExecute, to keep them up to date better.