Change the way compression editing works
When editing a length token, header token, or compressed token, always respect everything earlier in the format. Don't worry about respecting anything later in the format: you can throw it all away. Currently, we don't really expect anyone making changes like this except for a copy-paste scenario, so optimize for that use case. Clearing out remaining data and replacing it with blanks is the easiest way to limit unneeded repoints.