Fix broken test
This test was wrong. The final check verifiied that reading a pixel with a value "1" should return non-zero... even though that slot in the palette was black. Update it to make slot 1 of the palette non-black, and now the test works correctly.

The test actually passed before because of a bug where the color wasn't updating after sprite changes.