When adding reference anchors, make sure stuff points to them
The only reason we can find these is because we have a pointer to the table. So make sure that these pointers are properly recognized. They used to be, but it broke sometime when working on table expansion.
Related: When relocating a table for expansion, make sure there's enough room for all the extra elements, not just one.

Also, when checking for whether or not we need to seek for pointers, don't seek if the incoming run already has pointers attached.