Implement Home/End
If there's no data format, Home/End jump to the start/end of the line.
If there's a data format, Home/End jump to the start/ed of that format.

This is useful for quickly jumping to the end of a table, or selecting an entire table, or selecting an entire string/image/etc.