Add daycare exp multiplier
The daycare exp multiplier effects the step count, which is also used to control the timing for egg creation. Eggs will only be created when the step count is equal to 255 mod 256. Because of this, eggs will never be created if the step count increments by an even number. Any odd number is safe.

To help explain this to the user, a 'note' field has been added.