Fix a crash
When pasting specific data that was deep-copied, it's possible for a tilemap to temporarily see its tileset as another tilemap. This can get into a weird recursion case that can cause a stack overflow in some situations. The proper fix is to just not allow a tilemap to use another tilemap as a tileset. This was never a desired feature anyway: we just wanted to allow a tilemap to use a sprite as a tileset, to make discovery easier.