Add Help -> Update
The button is disabled if there is no new version available. The app will only check for updates once a day. The button will take you to the releases page so you can get the latest version - it won't update anything for you.

Also added a tooltip. Also adde a Discord link.