Performance Improvement: Table Tool Textboxes
TextBox is a lot slower to update than TextBlock. Since we have a whole lot of TextBoxes in the table, updating the table can randomly be very slow. To make this better, we'll use TextBox 'proxy' objects whenever the user isn't interacting with the TextBox: it's just a TextBlock that *looks* like a TextBox. The whole thing is wrapped in a UserControl for isolation. A single method in the ViewPort can swap between the two states.

We may later do something similar for the ComboBox, but it's fine for now.