Improve LZ rendering capabilities
Some programs create LZ compressed sprites that don't 'end' correctly. If those get inserted, the game will still be able to read them correctly, so I should be able to as well.
The new code allows for some sprites to express that they can 'AllowLengthErrors' and still work correctly. This allows this very specific type of read error to still get parsed correctly, but ONLY if it's in a table, and ONLY if it's a compressed sprite.