Improve UI when too many results
Part of this means that I need to limit the number of categories to something reasonable. But even when I do so, the number of unique elements in that category may be a bit crazy. Example: "stats/c" returns a lot of pokemon.
In those cases, just show the number of results and ask the user to filter further.