add minheight for goto view model
this is hacky: the ItemsPresenter within the ItemsControl is coming up with a default height of zero in some situations. I'm not sure how this is happening, but I can fix it (for now) by making sure that there's always at least 100 pixels of space for the buttons to display. As soon as any buttons get chosen, it seems to fix itself.