Add mail graphics table
the table includes not only the palette, tileset, and tilemap, but also the foreground/shadow colors for the text that goes on the mail. As such, tables need to be able to include elements that are uncompressed colors.

Add |c to specify that an element is a color (similar to |h). Add support for showing these new elements in the table tool.