Add metacommand for writing compressed lz data
If a user wants to cut/paste an existing run to freespace, the new format must be able to be put in the new spot. This means that as part of writing a format that does validity checks (like compressed data), it must be possible to first write 'empty' valid data to the free space.