Allow "beedrill" to go to "beedrill" even if 'names' is not the first item in the table list.
When the data for the current run changes (based on an address change), we update the table section and index.

When we update the table section or index, we update the address.

These two things can stomp on eachother if we're not careful. Introduce a flag so we can detect the recursion case.