Add support for selection movement in tilemaps
When moving blocks around within a tilemap, there's no easy way to guarantee that the user is moving in tile increments, and multilpe tiles may employ different palettes.

As such, conisder that moving a block around moves the *index* of the palette being used, not the color. This will mostly work as expected, but can cause colors to change as they enter or exit regions that use different palettes.